restructures, renames, updates highlighted components
